+++ packages/core/src/constants.ts
@@ -1,5 +1,12 @@
 import { defineChain } from "thirdweb";
-import { arbitrum, arbitrumSepolia, sepolia } from "thirdweb/chains";
+import {
+  arbitrum,
+  arbitrumSepolia,
+  base,
+  baseSepolia,
+  mainnet,
+  sepolia,
+} from "thirdweb/chains";
 
 import type { AddressString, Contract } from "./types";
 
@@ -75,10 +82,16 @@ export const MAGICSWAPV2_API_URL = {
     "https://api.goldsky.com/api/public/project_clrm53zqegpoi01x18coz2fb5/subgraphs/magicswapv2/live/gn",
   [arbitrumSepolia.id]:
     "https://api.goldsky.com/api/public/project_clrm53zqegpoi01x18coz2fb5/subgraphs/magicswapv2-dev/live/gn",
+  [treasure.id]:
+    "https://api.goldsky.com/api/public/project_clrm53zqegpoi01x18coz2fb5/subgraphs/magicswap-treasure/live/gn",
   [treasureTopaz.id]:
     "https://api.goldsky.com/api/public/project_clrm53zqegpoi01x18coz2fb5/subgraphs/magicswap-dev-topaz/live/gn",
   [sepolia.id]:
     "https://api.goldsky.com/api/public/project_clrm53zqegpoi01x18coz2fb5/subgraphs/magicswap-dev-sepolia/live/gn",
+  [base.id]:
+    "https://api.goldsky.com/api/public/project_clrm53zqegpoi01x18coz2fb5/subgraphs/magicswap-base/live/gn",
+  [baseSepolia.id]:
+    "https://api.goldsky.com/api/public/project_clrm53zqegpoi01x18coz2fb5/subgraphs/magicswap-dev-base-sepolia/live/gn",
 } as const;
 
 // Tokens
@@ -131,6 +144,11 @@ export const CONTRACT_ADDRESSES: Record<
     BulkTransferHelper: "0x0000000000c2d145a2526bd8c716263bfebe1a72",
     TreasureConduit: "0x74c4ed7e3bcbdf8e32b56d4d11d2cbd34ad4dc0b",
     TreasureBanners: "0x023083def4855918dea65f1dea351fcdabb1e073",
+  },
+  [mainnet.id]: {
+    // Tokens
+    MAGIC: "0xb0c7a3ba49c7a6eaba6cd4a96c55a1391070ac9a",
+    CRV: "0xD533a949740bb3306d119CC777fa900bA034cd52",
   },
   [sepolia.id]: {
     // Tokens